improve error handling for unrecognized mode
